+++ framework/CHANGELOG.md
@@ -13,7 +13,8 @@ Yii Framework 2 Change Log
 - Enh #12743: Added new methods `BaseActiveRecord::loadRelations()` and `BaseActiveRecord::loadRelationsFor()` to eager load related models for existing primary model instances (PowerGamer1)
 - Enh #20030: Improve performance of handling `ErrorHandler::$memoryReserveSize` (antonshevelev, rob006)
 - Enh #20042: Add empty array check to `ActiveQueryTrait::findWith()` (renkas)
-- Enh #20032: Added `mask` method for string masking with multibyte support (salehhashemi1992)
+- Enh #20032: Added `yii\helpers\BaseStringHelper::mask()` method for string masking with multibyte support (salehhashemi1992)
+- Enh #20034: Added `yii\helpers\BaseStringHelper::findBetween()` to retrieve a substring that lies between two strings (salehhashemi1992)
 
 
 2.0.49.2 October 12, 2023